Move shared functionality into a hook so we can easily pull that in where it's needed. Use this in the block controls regenerate button
